+++ src/main/java/roomescape/admin/controller/request/ReservationSearchRequest.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.admin.controller.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
