+++ src/main/java/roomescape/auth/web/controller/response/MemberNameResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member.controller.response;
+package roomescape.auth.web.controller.response;
 
 public record MemberNameResponse(String name) {
 }
