+++ src/main/java/roomescape/auth/web/resolver/AuthenticatedMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.member.resolver;
+package roomescape.auth.web.resolver;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -10,7 +10,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.member.infrastructure.jwt.JwtHandler;
+import roomescape.auth.infrastructure.jwt.JwtHandler;
 
 @RequiredArgsConstructor
 @Component
