+++ src/main/java/roomescape/config/web/WebMvcConfig.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.member.interceptor.AdminMemberHandlerInterceptor;
-import roomescape.member.resolver.AuthenticatedMemberArgumentResolver;
+import roomescape.auth.web.interceptor.AdminMemberHandlerInterceptor;
+import roomescape.auth.web.resolver.AuthenticatedMemberArgumentResolver;
 
 @RequiredArgsConstructor
 @Configuration
